Finalizado o Projeto

+++ src/main/java/com/zup/orangeTalents/repository/PessoaRepository.java
@@ -3,7 +3,7 @@ package com.zup.orangeTalents.repository;
 import com.zup.orangeTalents.entity.Pessoa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PessoaRepository extends JpaRepository<Pessoa, Integer> {
+public interface PessoaRepository extends JpaRepository<Pessoa, Long> {
     Pessoa findByNome(String name);
     Pessoa findByCpf(String cpf);
 }
